Fix inaccurate port of Linux kernel cmpxchg()
Linux uses a macro, which returns the type of operands. In my
implementation I used inline function, which always returned long.
If I get an integer with the most significant bit set, I got the bit
extension on implicit type conversion and the operation always treated
as failed. This commit just reuses the original Linux macro.