The new data management architecture
This is a dirty rework of the new architecture with many 'AK'-marked
comments, which must be addressed to make the commit work. Besides HTrie
it addresses tasks:
* https://github.com/tempesta-tech/tempesta/issues/515
* https://github.com/tempesta-tech/tempesta/issues/516
* https://github.com/tempesta-tech/tempesta/issues/400

Tables and table spaces were introduced for flexible table sizes.

Block allocation is in progress of full rework: there are no more 4KB block
allocations and we move with global, tables shared, extents allocation
and simple segregated lists per-cpu per-table heaps.

External memory fragmentation will be addressed with eviction, so
evict.c was introduced with the bunch of TODOs.

Now TDB will be able to return large allocations up to 2MB.